Rework math for P173
Now we compute the number of laminae by fixing the number of layers.

Since the number of layers is capped at a much smaller number (it scales
as the square root of the max number of tiles) the algorithm changes
from O(T) to O(Sqrt[T]). This is irrelevant for the 10^6 case provided,
but now we can solve much bigger numbers (10^12 for instance).